bug: fix client side Date issue server component Date was out of sync with client date on browser. Split into client component to fix issue
